+++ src/main/main.js
@@ -357,8 +357,6 @@ function initIpcMainHandle(ipcMain) {
   });
 
   ipcMain.handle("store:modifyNote", function(event, note) {
-    log.info("store:modifyNote", note, n3.repository);
-
     return n3.repository.modifyNote(note);
   });